0.54.0
Deprecation Notices
Using an underscore in a flag name is deprecated and has been replaced by a hyphen. The underscore format will remain compatible for some time, until it is completely removed in a future version. For example, --remote_port is replaced with --remote-port.
Features
The Refresh and ClearOfflineProxies buttons have been added to the Dashboard of frps.
Fixes
The host/domain matching in the routing rules has been changed to be case-insensitive.